# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: EvalFrog External API
  version: 1.0.0-m10
  description: Human Web and Agent CLI share this complete authoring and runtime command API.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /v1/projects/{project_id}/workflows:
    post:
      operationId: createWorkflow
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkflowRequest' }
      responses:
        '201': { $ref: '#/components/responses/WorkflowCreated' }
        default: { $ref: '#/components/responses/Error' }
  '/v1/projects/{project_id}/workflows:copy':
    post:
      operationId: copyPublishedVersion
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CopyWorkflowRequest' }
      responses:
        '201': { $ref: '#/components/responses/WorkflowCreated' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/workflows/{workflow_id}:
    get:
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
      responses:
        '200':
          description: Project-scoped Workflow metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workflow' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/workflows/{workflow_id}/draft:
    get:
      operationId: getDraft
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
      responses:
        '200':
          description: Current Draft pointer and immutable revision.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Draft' }
        default: { $ref: '#/components/responses/Error' }
    put:
      operationId: saveDraft
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveDraftRequest' }
      responses:
        '201':
          description: Newly created immutable Draft Revision.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftRevision' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/workflows/{workflow_id}/draft/validate:
    post:
      operationId: validateDraft
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevisionRequest' }
      responses:
        '200':
          description: Stable ordered diagnostics; valid can be false.
          content:
            application/json:
              schema:
                type: object
                required: [valid, diagnostics]
                additionalProperties: false
                properties:
                  valid: { type: boolean }
                  diagnostics: { type: array, items: { type: object } }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/workflows/{workflow_id}/draft/test:
    post:
      operationId: testDraft
      description: Server validates and compiles the selected immutable Draft Revision, then creates a Test Run bound to the resulting Snapshot.
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestDraftRequest' }
      responses:
        '201':
          description: Persisted Test Run. The immutable Snapshot remains server-owned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunRecord' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/workflows/{workflow_id}/publish:
    post:
      operationId: publishWorkflow
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishRequest' }
      responses:
        '201':
          description: New immutable Version, automatically activated.
          content:
            application/json:
              schema:
                type: object
                required: [version]
                additionalProperties: false
                properties:
                  version: { $ref: '#/components/schemas/PublishedVersion' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/workflows/{workflow_id}/versions/{version_number}/activate:
    post:
      operationId: rollbackWorkflow
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
        - name: version_number
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        '200':
          description: Historical immutable Version selected as Active.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishedVersion' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/workflows/{workflow_id}/runs:
    post:
      operationId: createProductionRun
      description: Creates a Run from the active immutable Published Version only.
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/WorkflowID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRunRequest' }
      responses:
        '201':
          description: Persisted Production Run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunRecord' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/runs/{run_id}:
    get:
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/RunID'
      responses:
        '200':
          description: Current PostgreSQL-derived Run Read Model.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunView' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/runs/{run_id}/events:
    get:
      operationId: subscribeRunUpdates
      description: Lossy Server-Sent Event wake-up stream. Consumers must re-read the Run view after every update or reconnect.
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/RunID'
      responses:
        '200':
          description: Run update wake-up events only; no Runtime state is carried in the stream.
          content:
            text/event-stream:
              schema: { type: string }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/ProjectID'
        - $ref: '#/components/parameters/RunID'
      responses:
        '202': { $ref: '#/components/responses/CancelAccepted' }
        '200': { $ref: '#/components/responses/CancelAccepted' }
        default: { $ref: '#/components/responses/Error' }
  /v1/node-types:
    get:
      operationId: listNodeTypes
      responses:
        '200':
          description: Public, versionless authoring descriptions from the current catalog.
          content:
            application/json:
              schema:
                type: object
                required: [node_types]
                properties:
                  node_types:
                    type: array
                    items: { $ref: '#/components/schemas/NodeDescription' }
        default: { $ref: '#/components/responses/Error' }
  /v1/projects/{project_id}/connections:
    get:
      operationId: listConnections
      parameters:
        - $ref: '#/components/parameters/ProjectID'
      responses:
        '200':
          description: Safe authoring metadata; endpoints and secret references are absent.
          content:
            application/json:
              schema:
                type: object
                required: [connections]
                properties:
                  connections:
                    type: array
                    items: { $ref: '#/components/schemas/ConnectionSummary' }
        default: { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    WorkflowID:
      name: workflow_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    RunID:
      name: run_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
  schemas:
...
    CreateWorkflowRequest:
      type: object
      required: [name, ir]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        ir: { $ref: '#/components/schemas/IRDocument' }
    SaveDraftRequest:
      type: object
      required: [expected_revision, ir]
      additionalProperties: false
      properties:
        expected_revision: { type: integer, format: int64, minimum: 1 }
        ir: { $ref: '#/components/schemas/IRDocument' }
    RevisionRequest:
      type: object
      required: [revision]
      additionalProperties: false
      properties:
        revision: { type: integer, format: int64, minimum: 1 }
    TestDraftRequest:
      type: object
      required: [revision, input, deadline_at]
      additionalProperties: false
      properties:
        revision: { type: integer, format: int64, minimum: 1 }
        input: { type: object }
        deadline_at: { type: string, format: date-time }
    CreateRunRequest:
      type: object
      required: [input, deadline_at]
      additionalProperties: false
      properties:
        input: { type: object }
        deadline_at: { type: string, format: date-time }
    PublishRequest:
      type: object
      required: [expected_revision, change_log]
      additionalProperties: false
      properties:
        expected_revision: { type: integer, format: int64, minimum: 1 }
        change_log: { type: string, maxLength: 4000 }
    CopyWorkflowRequest:
      type: object
      required: [source_workflow_id, source_version_number, name]
      additionalProperties: false
      properties:
        source_workflow_id: { type: string, format: uuid }
        source_version_number: { type: integer, format: int64, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 200 }
    Workflow:
      type: object
      required: [workflow_id, project_id, name, created_by, created_at, updated_at]
      properties:
        workflow_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string }
        active_version_id: { type: [string, 'null'], format: uuid }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DraftRevision:
      type: object
      required: [draft_revision_id, project_id, workflow_id, revision_number, ir, catalog_revision, created_by, created_at]
      properties:
        draft_revision_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        workflow_id: { type: string, format: uuid }
        revision_number: { type: integer, format: int64 }
        ir: { $ref: '#/components/schemas/IRDocument' }
        catalog_revision: { type: string }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        cloned_from_version_id: { type: [string, 'null'], format: uuid }
    Draft:
      type: object
      required: [project_id, workflow_id, current_revision, state_version, current]
      properties:
        project_id: { type: string, format: uuid }
        workflow_id: { type: string, format: uuid }
        current_revision: { type: integer, format: int64 }
        state_version: { type: integer, format: int64 }
        current: { $ref: '#/components/schemas/DraftRevision' }
...
    RunRecord:
      type: object
      description: Minimal command response. Read current Runtime state through the Run View endpoint.
      required: [run_id, purpose, state, created_at]
      properties:
        run_id: { type: string, format: uuid }
        purpose: { type: string, enum: [test, production] }
        created_at: { type: string, format: date-time }
        state: { type: string, enum: [pending, running, succeeded, failed, canceled, timed_out] }
...
  responses:
    WorkflowCreated:
      description: Workflow and immutable Draft Revision 1.
      content:
        application/json:
          schema:
            type: object
            required: [workflow, draft_revision]
            additionalProperties: false
            properties:
              workflow: { $ref: '#/components/schemas/Workflow' }
              draft_revision: { $ref: '#/components/schemas/DraftRevision' }
...
    Error:
      description: Stable machine-readable error envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }